--- src/main/java/com/example/RaiseTechTask7remake/DeleteForm.java
@@ -1,18 +1,0 @@
-package com.example.RaiseTechTask7remake;
-
-public class DeleteForm {
-    private String name;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
-
-
-
-
-
