+++ src/main/java/com/example/Raisetechtask7remake/RaiseTechTask7RemakeApplication.java
@@ -1,4 +1,4 @@
-package com.example.RaiseTechTask7remake;
+package com.example.Raisetechtask7remake;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
