+++ src/test/java/com/example/Raisetechtask7remake/RaiseTechTask7RemakeApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.RaiseTechTask7remake;
+package com.example.Raisetechtask7remake;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
